+++ CHANGELOG.md
@@ -58,3 +58,4 @@ Changed repo name from brainspin to cvasl
 
     - command-line sex recoding over a folder
     - command-line loging of columns over a folder
+    - command-line cleaning off unwanted columns run over a folder